Revert "Modularize and Add Timers"
This reverts commit 810e7b000071d0ac5d0e57e3f2dac17b4cb69cb1.

+++ HADES_Software/tm4c_fatfs.c
@@ -21,7 +21,6 @@
 #include "driverlib/sysctl.h"
 
 #include "fatfs/diskio.h"
-#include "fatfs/integer.h"
 
 #include "userlib/spim_drv.h"
 
@@ -459,7 +458,7 @@ DSTATUS disk_status(BYTE drv /* Physical drive number (0) */
 DRESULT disk_read(BYTE drv, /* Physical drive number (0) */
                   BYTE *buff, /* Pointer to the data buffer to store read data */
                   DWORD sector, /* Start sector number (LBA) */
-                  UINT count /* Sector count (1..255) */
+                  BYTE count /* Sector count (1..255) */
                   ) {
   if (drv || !count) return RES_PARERR;
   if (Stat & STA_NOINIT) return RES_NOTRDY;
@@ -495,7 +494,7 @@ DRESULT disk_read(BYTE drv, /* Physical drive number (0) */
 DRESULT disk_write(BYTE drv, /* Physical drive number (0) */
                    const BYTE *buff, /* Pointer to the data to be written */
                    DWORD sector, /* Start sector number (LBA) */
-                   UINT count /* Sector count (1..255) */
+                   BYTE count /* Sector count (1..255) */
                    ) {
   if (drv || !count) return RES_PARERR;
   if (Stat & STA_NOINIT) return RES_NOTRDY;
